fix: gracefully handle cases where content URIs are invalid
Prior to this commit, all content URIs of application versions are
expected to follow the format `<provider>:<identifier>`. In case a
content URI doesn't follow this format, an error is thrown.

We now ensure that in such cases, the code still executes but resolves
early with an error or warning respectively.

It also fixes #27 where parsing content URIs that contain more than
one split character (e.g. `:` in `http://localhost:8080/foo/bar`) splits
the `identifier` as well (which we don't want).

Fixes #20, #27